fix(#8): List 컴포넌트로 넘겨주는 list 데이터 중복 로직 최소화
- List 컴포넌트를 ToDoList 컴포넌트로 변경하고 해당 컴포넌트에 관련된 비즈니스 로직 이동
- 외부에서는 type만 받아 list 종류만 명시하도록 구현
